pplay now includes sxyca directly so it's not necessary to install pplay pip package
